Compute a list of canonical covariance matrices

Compute a list of canonical covariance matrices

create_cov_canonical(R, singletons = T, hetgrid = c(0, 0.25, 0.5, 0.75,
  1))

Arguments

R
an integer indicating the number of conditions

singletons
a logical value indicating whether the singleton matrices are computed

hetgrid
a vector of numbers between -1 and 1, each representing the off-diagonal elements of matrices with 1s on the diagonal. If 0 is included, the identity matrix will be returned which corresponds to assuming effects are independent across conditions. IF NULL, these matrices are not computed.

Value

a list of canonical covariance matrices

Details

This function computes canonical covariance matrices to be provided to mash

Examples

 mmbr:::create_cov_canonical(3)
 mmbr:::create_cov_canonical(3, singletons=F)
 mmbr:::create_cov_canonical(3, hetgrid=NULL)