Create mash prior object

Create mash prior object

create_mash_prior(
  fitted_g = NULL,
  mixture_prior = NULL,
  sample_data = NULL,
  null_weight = NULL,
  use_grid = FALSE,
  weights_tol = 1e-10,
  max_mixture_len = 40,
  include_indices = NULL,
  ...
)

Arguments

fitted_g
from mashr::mash

mixture_prior
a list of (weights = vector(), matrices = list()) where matrices is a list of prior matrices and have same length as weights.

sample_data
a list of (X=X,Y=Y,residual_variance=residual_variance,center=T,scale=T) to allow for automatically determine canonical priors with equal weights

null_weight
whether or not to add a weight for null in single effect models. By default it takes the null weight from fitted_g if available. Use `null_weight = 0` to override the behavior.

use_grid
expand mixture by grid values as in MASH (not necessary when prior scalar is estimated)

weights_tol
filter out priors with weights smaller than weights_tol

max_mixture_len
only keep the top priors by weight so that the list of mixture prior is of max_mixture_len. Use `max_mixture_len=-1` to include all input weights after weights_tol filtering. Default is set to length 40.

include_indices
postprocess input prior to only include conditions from this indices

...
other parameters, for mmbr:::create_cov_canonical

Value

mash prior object for use with msusie() function

Details

...