SUm of Single Effect (SuSiE) Regression of multivariate Y on X

SUm of Single Effect (SuSiE) Regression of multivariate Y on X

msusie(
  X,
  Y,
  L = 10,
  prior_variance = 0.2,
  residual_variance = NULL,
  prior_weights = NULL,
  standardize = TRUE,
  intercept = TRUE,
  estimate_residual_variance = FALSE,
  estimate_prior_variance = TRUE,
  estimate_prior_method = "EM",
  check_null_threshold = 0,
  prior_tol = 1e-09,
  compute_objective = FALSE,
  s_init = NULL,
  coverage = 0.95,
  min_abs_corr = 0.5,
  compute_univariate_zscore = FALSE,
  precompute_covariances = FALSE,
  max_iter = 100,
  tol = 0.001,
  verbose = TRUE,
  track_fit = FALSE
)

Arguments

X
an n by p matrix of covariates

Y
an n vector, or n by r matrix of response variables

L
maximum number of non-zero effects

prior_variance
Can be 1) a vector of length L, or a scalar, for scaled prior variance when Y is univariate (equivalent to `susieR::susie`); 2) a matrix for simple Multivariate regression or 3) a MASH fit that contains an array of prior covariance matrices and their weights

residual_variance
the residual variance (defaults to sample variance of Y)

prior_weights
a p vector of prior probability that each element is non-zero

standardize
logical flag (default=TRUE) for whether to standardize columns of X to unit variance prior to fitting. If you do not standardize you may need to think more carefully about specifying the scale of prior variance. Whatever the value of standardize, the coefficients (returned from `coef`) are for X on the original input scale. Any column of X that has zero variance is not standardized, but left as is.

intercept
Should intercept be fitted (default=TRUE) or set to zero (FALSE). The latter is generally not recommended.

estimate_residual_variance
indicates whether to estimate residual variance (currently only works for univariate Y input)

estimate_prior_variance
indicates whether to estimate prior (currently only works for univariate Y and for multivariate Y when prior is a single matrix)

estimate_prior_method
the method used for estimating prior variance: "optim", "uniroot" and "em" for univariate Y, "optim" and "simple" for multivariate Y.

check_null_threshold
when prior variance is estimated, compare the estimate with the null and set prior variance to null (zero) unless the log-likelihood using the estimate is larger than that of null by this threshold. For example, you can set it to 0.1 to nudge the estimate towards zero. When used with "EM" method setting check_null_threshold=NA will skip the check and instead relying solely on EM to update this parameter.

prior_tol
when prior variance is estimated, compare the estimated value to this tol at the end of the analysis and exclude a single effect from PIP computation if the estimated prior variance is smaller than it.

s_init
a previous susie fit with which to initialize

coverage
coverage of confident sets. Default to 0.95 for 95% credible interval.

min_abs_corr
minimum of absolute value of correlation allowed in a credible set. Default set to 0.5 to correspond to squared correlation of 0.25, a commonly used threshold for genotype data in genetics studies.

compute_univariate_zscore
if TRUE outputs z-score from per variable univariate regression

precompute_covariances
if TRUE, precomputes various covariance quantities to speed up computations at the cost of increased memory usage

max_iter
maximum number of iterations to perform

tol
convergence tolerance

verbose
if TRUE outputs some progress messages

track_fit
add an attribute trace to output that saves some current quantities of all iterations

Value

a susie fit, which is a list with some or all of the following elements

alpha

an L by p matrix of posterior inclusion probabilites

b1

an L by p matrix of posterior means (conditional on inclusion)

b2

an L by p matrix of posterior second moments (conditional on inclusion)

KL

an L vector of KL divergence

lbf

an L vector of logBF

sigma2

residual variance

V

prior variance

elbo

a vector of values of elbo achieved (objective function)

niter

number of iterations took for convergence

convergence

convergence status

sets

a list of `cs`, `purity` and selected `cs_index`

pip

a vector of posterior inclusion probability

walltime

records runtime of the fitting algorithm

z

a vector of univariate z-scores

Details

Performs Bayesian multiple linear regression of Y on X. That is, this function fits the regression model Y= sum_l Xb_l + e, where elements of e are iid N(0,residual_variance) and the sum_l b_l is a p vector of effects to be estimated. The assumption is that each b_l has exactly one non-zero element, with all elements equally likely to be non-zero.

Examples

set.seed(1)
n = 1000
p = 1000
beta = rep(0,p)
beta[1:4] = 1
X = matrix(rnorm(n*p),nrow=n,ncol=p)
y = X %*% beta + rnorm(n)
res = msusie(X,y,L=10)