SUm of Single Effect (SuSiE) Regression using Summary Statistics Z and R

SUm of Single Effect (SuSiE) Regression using Summary Statistics Z and R

msusie_rss(
  Z,
  R,
  L = 10,
  r_tol = 1e-08,
  prior_variance = 50,
  residual_variance = NULL,
  prior_weights = NULL,
  estimate_residual_variance = FALSE,
  estimate_prior_variance = TRUE,
  estimate_prior_method = "simple",
  compute_objective = FALSE,
  precompute_covariances = FALSE,
  s_init = NULL,
  coverage = 0.95,
  min_abs_corr = 0.5,
  max_iter = 100,
  tol = 0.001,
  z_thresh = 2,
  verbose = TRUE,
  track_fit = FALSE
)

Arguments

Z
a J by R matrix of z scores

R
a J by J LD matrix

L
maximum number of non-zero effects

prior_variance
Can be 1) a vector of length L, or a scalar, for scaled prior variance when Y is univariate (equivalent to `susieR::susie`); 2) a matrix for simple Multivariate regression or 3) a MASH fit that contains an array of prior covariance matrices and their weights

residual_variance
the residual variance (defaults to 1)

prior_weights
a p vector of prior probability that each element is non-zero

estimate_residual_variance
indicates whether to estimate residual variance (currently only works for univariate Y input)

estimate_prior_variance
indicates whether to estimate prior (currently only works for univariate Y and for multivariate Y when prior is a single matrix)

estimate_prior_method
the method used for estimating prior variance: "optim", "uniroot" and "em" for univariate Y, "optim" and "simple" for multivariate Y.

s_init
a previous susie fit with which to initialize

coverage
coverage of confident sets. Default to 0.95 for 95% credible interval.

min_abs_corr
minimum of absolute value of correlation allowed in a credible set. Default set to 0.5 to correspond to squared correlation of 0.25, a commonly used threshold for genotype data in genetics studies.

max_iter
maximum number of iterations to perform

tol
convergence tolerance

z_thresh
the z score threshold below which to call an effect null

verbose
if TRUE outputs some progress messages

track_fit
add an attribute trace to output that saves some current quantities of all iterations

Value

a susie fit, which is a list with some or all of the following elements

alpha

an L by p matrix of posterior inclusion probabilites

b1

an L by p matrix of posterior means (conditional on inclusion)

b2

an L by p matrix of posterior second moments (conditional on inclusion)

KL

an L vector of KL divergence

lbf

an L vector of logBF

sigma2

residual variance

V

prior variance

elbo

a vector of values of elbo achieved (objective function)

niter

number of iterations took for convergence

convergence

convergence status

sets

a list of `cs`, `purity` and selected `cs_index`

pip

a vector of posterior inclusion probability

walltime

records runtime of the fitting algorithm

z

a vector of univariate z-scores

Examples

set.seed(1)
n = 1000
p = 1000
beta = rep(0,p)
beta[1:4] = 1
X = matrix(rnorm(n*p),nrow=n,ncol=p)
y = X %*% beta + rnorm(n)
R = t(X) %*% X
z = susieR:::calc_z(X,y)
res = msusie_rss(z,R,L=10)